refactor: independent join page and dropdown
