+++ web/app/habit/page.tsx
@@ -9,13 +9,9 @@ import Onboard from './components/Onboard';
 import Dashboard from './components/Dashboard';
 
 
-const nouns = require('../../src/imgs/nouns.png') as string;
-
-import { useAccount } from 'wagmi';
+const nouns = require('@/imgs/nouns.png') as string;
 
 export default function DashboardPage() {
-  
-  // const { address } = useAccount();
 
   const address = '0xBAbe69e7F2C7A9f0369Ae934865d0097B73543Fc'
