+++ web/app/habit/stake/page.tsx
@@ -4,48 +4,41 @@
 'use client';
 
 import Image from 'next/image';
-import { useRef, useState, SetStateAction, useEffect } from 'react';
+import { useRef, useState, useEffect } from 'react';
 import { useAccount, useBalance, useWriteContract, useWaitForTransactionReceipt } from 'wagmi';
 import crypto from 'crypto';
 import { GateFiDisplayModeEnum, GateFiSDK, GateFiLangEnum } from '@gatefi/js-sdk';
 import { parseEther } from 'ethers/lib/utils';
 import trackerContract from '@/contracts/tracker.json';
-import toast from 'react-hot-toast';
+import toast, { Toaster } from 'react-hot-toast';
 
-import { redirect } from 'next/navigation';
 import { Challenge } from '@/hooks/useUserChallenges';
 
 const img = require('@/imgs/step2.png') as string;
 const kangaroo = require('@/imgs/kangaroo.png') as string;
 
-import { challenges } from '@/constants';
 import { formatEther } from 'viem';
 
-import { ChallengesDropDown } from './dropdown';
+import { ChallengesDropDown } from './components/dropdown';
+import { challenges } from '@/constants';
 
-export default function Join({
-  setSteps,
-  setSelectedChallenge,
-  selectedChallenge,
-}: {
-  setSteps: React.Dispatch<SetStateAction<number>>;
-  setSelectedChallenge: React.Dispatch<SetStateAction<Challenge>>;
-  selectedChallenge: Challenge;
-}) {
+const nouns = require('@/imgs/nouns.png') as string;
+
+export default function Join() {
   const overlayInstanceSDK = useRef<GateFiSDK | null>(null);
   const [isOverlayVisible, setIsOverlayVisible] = useState(false);
 
+  const [selectedChallenge, setSelectedChallenge] = useState<null | Challenge>(null);
+
   const handleOnChoose = (challengeArx: string) => {
-    const challenge = challenges.find((c) => c.arxAddress === challengeArx);
-    if (challenge) {
-      setSelectedChallenge(challenge);
-    }
+    const challenge = challenges.find(c => c.arxAddress === challengeArx)?? null
+    setSelectedChallenge(challenge)
   };
 
   const { address: smartWallet } = useAccount();
   const balance = useBalance({ address: smartWallet });
   const ethBalance = balance.data ? Number(formatEther(balance.data.value)) : 0;
-  const hasEnoughBalance = ethBalance > selectedChallenge.stake;
+  const hasEnoughBalance = selectedChallenge && ethBalance > selectedChallenge.stake;
 
   const handleOnClickOnramp = async () => {
     if (overlayInstanceSDK.current) {
@@ -105,6 +98,7 @@ export default function Join({
   });
 
   const onJoinButtonClick = async () => {
+    if (!selectedChallenge) return;
     writeContract({
       address: trackerContract.address as `0x${string}`,
       abi: trackerContract.abi,
@@ -114,9 +108,9 @@ export default function Join({
     });
   };
 
-  if (!smartWallet) {
-    redirect('/');
-  }
+  // if (!smartWallet) {
+  //   redirect('/');
+  // }
 
   useEffect(() => {
     if (isSuccess) {
@@ -124,7 +118,6 @@ export default function Join({
       setTimeout(() => {
         // setSteps(3);
         // todo: go to checkin page, or dashboard
-
       }, 2000);
     }
   }, [isSuccess]);
@@ -136,60 +129,71 @@ export default function Join({
   }, [joinError]);
 
   return (
-    <div className="flex flex-col items-center justify-center">
-      {/* Img and Description */}
-      <div className="flex items-center gap-6">
-        <Image src={img} width="50" alt="Step 2 Image" className="mb-3 rounded-full object-cover" />
-        <p className="mr-auto text-lg font-bold">Stake and join a challenge</p>
-      </div>
-      <div>
-        <p className="p-4 text-sm"> Choose Challenge </p>
+    <main className="container mx-auto flex flex-col items-center px-8 pt-16">
+      <Toaster />
+      <Image src={nouns} width="100" height="100" alt="Nouns Logo" className="mb-10" />
+      <div className="font-title container mb-10 w-full text-center text-4xl">
+        {' '}
+        Alibuda Habit Builder{' '}
       </div>
 
-      {/* drop down here */}
-      <div>
-        <ChallengesDropDown
-          selectedChallenge={selectedChallenge}
-          setSelectedChallenge={setSelectedChallenge}
-          onChoose={handleOnChoose}
-        />
-      </div>
+      <div className="flex flex-col items-center justify-center">
+        {/* Img and Description */}
+        <div className="flex items-center gap-6">
+          <Image
+            src={img}
+            width="50"
+            alt="Step 2 Image"
+            className="mb-3 rounded-full object-cover"
+          />
+          <p className="mr-auto text-lg font-bold">Stake and Join a challenge</p>
+        </div>
+        
+        {/* drop down here */}
+        <div className='pt-4'>
+          <ChallengesDropDown
+            selectedChallenge={selectedChallenge}
+            setSelectedChallenge={setSelectedChallenge}
+            onChoose={handleOnChoose}
+          />
+        </div>
 
-      <button
-        type="button"
-        className="bg-yellow bg-yellow mt-4 rounded-lg border-solid px-6 py-3 font-bold disabled:cursor-not-allowed disabled:bg-gray-400"
-        style={{ width: '250px', height: '45px', color: 'white' }}
-        onClick={onJoinButtonClick}
-        disabled={!hasEnoughBalance || joinPending || isLoading}
-      >
-        Stake {selectedChallenge.stake} ETH
-      </button>
+        <button
+          type="button"
+          className="bg-yellow bg-yellow mt-4 rounded-lg border-solid px-6 py-3 font-bold disabled:cursor-not-allowed disabled:bg-gray-400"
+          style={{ width: '250px', height: '45px', color: 'white' }}
+          onClick={onJoinButtonClick}
+          disabled={!hasEnoughBalance || joinPending || isLoading}
+        >
+          {selectedChallenge ? `Stake ${selectedChallenge.stake} ETH` : 'Stake'}
+        </button>
 
-      <div className="p-4 text-xs">
-        {balance.data && hasEnoughBalance ? (
-          <p> 💰 Smart Wallet Balance: {ethBalance.toString()} ETH </p>
-        ) : (
-          <p>
-            {' '}
-            🚨 Insufficient Balance: {ethBalance.toString()} ETH.{' '}
-            <span className="font-bold hover:underline" onClick={handleOnClickOnramp}>
+        <div className="p-4 text-xs">
+          {balance.data && hasEnoughBalance ? (
+            <p> 💰 Smart Wallet Balance: {ethBalance.toString()} ETH </p>
+          ) : (
+            <p>
               {' '}
-              Onramp now{' '}
-            </span>{' '}
-          </p>
-        )}
-      </div>
-      <div id="overlay-button"> </div>
+              🚨 Insufficient Balance: {ethBalance.toString()} ETH.{' '}
+              <span className="font-bold hover:underline" onClick={handleOnClickOnramp}>
+                {' '}
+                Onramp now{' '}
+              </span>{' '}
+            </p>
+          )}
+        </div>
+        <div id="overlay-button"> </div>
 
-      {/* warn message */}
-      <div className="text-md px-10 pt-8">
-        if you fail to maintain the habit, 50% of the stake will be donated to designated public
-        goods orgs, and 50% be distributed to other habit building winners.
-      </div>
+        {/* warn message */}
+        <div className="text-md px-10 pt-8">
+          if you fail to maintain the habit, 50% of the stake will be donated to designated public
+          goods orgs, and 50% be distributed to other habit building winners.
+        </div>
 
-      <div className="w-full justify-start px-6">
-        <Image src={kangaroo} width="350" alt="Kangaroo" className="mb object-cover" />
+        <div className="w-full justify-start px-6">
+          <Image src={kangaroo} width="350" alt="Kangaroo" className="mb object-cover" />
+        </div>
       </div>
-    </div>
+    </main>
   );
 }
